Changeling armor qdel check
Because if it got deleted, it'd force you to drop it...Which would qdel it again.

+++ code/game/gamemodes/changeling/powers/armor.dm
@@ -50,7 +50,8 @@
 
 /obj/item/clothing/suit/space/changeling/dropped(mob/user)
 	..()
-	qdel(src)
+	if(!QDELETED)
+		qdel(src)
 
 /obj/item/clothing/head/helmet/space/changeling
 	name = "flesh mass"
